projects/workflow_v_agent/python_version/pyproject.toml
[project]
name = "workflow-v-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "colorama>=0.4.6",
    "google-genai>=1.2.0",
    "python-dotenv>=1.0.0",
]
